Forbid pass non-node object to graphql printer
