+++ src/language-js/printer-estree.js
@@ -12,6 +12,7 @@ import {
   indent,
 } from "../document/builders.js";
 import { replaceEndOfLine } from "../document/utils.js";
+import createPrintPreCheckFunction from "../utils/create-print-pre-check-function.js";
 import embed from "./embed.js";
 import clean from "./clean.js";
 import { insertPragma } from "./pragma.js";
@@ -87,6 +88,8 @@ import { printBlock, printBlockBody } from "./print/block.js";
 import { printComment } from "./print/comment.js";
 import { printLiteral } from "./print/literal.js";
 import { printDecorators } from "./print/decorators.js";
+
+const ensurePrintingNode = createPrintPreCheckFunction(getVisitorKeys);
 
 function genericPrint(path, options, print, args) {
   const node = path.getValue();
@@ -836,46 +839,6 @@ function canAttachComment(node) {
     // `babel-ts` don't have similar node for `class Foo { bar() /* bat */; }`
     node.type !== "TSEmptyBodyFunctionExpression"
   );
-}
-
-function ensurePrintingNode(path) {
-  let name = path.getName();
-
-  // AST root
-  if (name === null) {
-    return;
-  }
-
-  if (typeof name === "number") {
-    /*
-    Nodes in array are stored in path.stack like
-
-    ```js
-    [
-      parentNode,
-      property, // <-
-      array,
-      index,
-      node,
-    ]
-    ```
-    */
-    name = path.stack[path.stack.length - 4];
-  }
-
-  const parent = path.getParentNode();
-  const visitorKeys = getVisitorKeys(parent);
-  if (visitorKeys.includes(name)) {
-    return;
-  }
-
-  /* istanbul ignore next */
-  throw Object.assign(new Error("Calling `print()` on non-node object."), {
-    parentNode: parent,
-    allowedProperties: visitorKeys,
-    printingProperty: name,
-    printingValue: path.getValue(),
-  });
 }
 
 const printer = {
